Add a visualization pane and a theme generator (#3)
The visualization pane is a simple display for a widget tree
The theme generator generates widget trees for display in the visualization pane
Currently a rudimentary version that displays some lobby elements are in effect, with appropriate to-dos provided.
Also some refactoring of the code to have better design principles.